# Speed up build of Python bindings