--- a/python/common/qgis_sip_project.py
+++ b/python/common/qgis_sip_project.py
@@ -1,3 +1,18 @@
+# ***************************************************************************
+#     qgis_sip_project.py
+#     ---------------------
+#     Date                 : September 2021
+#     Copyright            : (C) 2021 by Sandro Mani
+#     Email                : manisandro at gmail dot com
+# ***************************************************************************
+# *                                                                         *
+# *   This program is free software; you can redistribute it and/or modify  *
+# *   it under the terms of the GNU General Public License as published by  *
+# *   the Free Software Foundation; either version 2 of the License, or     *
+# *   (at your option) any later version.                                   *
+# *                                                                         *
+# ***************************************************************************
+
 from pathlib import Path
 import shutil
 from pyqtbuild import PyQtBindings, PyQtProject