## Description

As my work on [incremental builds of Python bindings](https://github.com/qgis/QGIS-Enhancement-Proposals/pull/338) is taking longer than anticipated, I thought it best to publish some of my changes that will improve clean build times in the meantime. This PR, along with another [PR to SIP](https://github.com/Python-SIP/sip/pull/87), should improve the performance of building `qgis._core` about two times, with larger improvements for simpler changes.

Here are some of my measurements of build times (`ninja _core.so`), with "after" measured after both this PR and the SIP PR have been applied:

| Description         | Before | After |
|---------------------|--------|-------|
| Clean build of core | 393 s  | 187 s |
| Docstring change    | 389 s  | 74 s  |
| New method          | 396 s  | 149 s |

You can install my branch of SIP with `pip3 install git+https://github.com/dvdkon/sip.git@speedup`

The main change on the QGIS side, apart from some refactoring of previously duplicated code, is adding a build step that only modifies those `.cpp` files that have been modified. This means that the modification timestamps of other files are preserved, so Ninja will only rebuild what is necessary.

<!--
  BEFORE HITTING SUBMIT -- Please BUILD AND TEST your changes thoroughly. This is YOUR responsibility! Do NOT rely on the QGIS code maintainers to do this for you!!

  IMPORTANT NOTES FOR FIRST TIME CONTRIBUTORS
  ===========================================

  Congratulations, you are about to make a pull request to QGIS! To make this as easy and pleasurable for everyone, please take the time to read these lines before opening the pull request.

  Include a few sentences describing the overall goals for this pull request (PR). If applicable also add screenshots or - even better - screencasts.
  Include both: *what* you changed and *why* you changed it.

  If this is a pull request that adds new functionality which needs documentation, give an especially detailed explanation.
  In this case, start with a short abstract and then write some text that can be copied 1:1 to the documentation in the best case.

  Also mention if you think this PR needs to be backported. And list relevant or fixed issues.

------------------------

  Reviewing is a process done by project maintainers, mostly on a volunteer basis. We try to keep the overhead as small as possible and appreciate if you help us to do so by checking the following list.
  Feel free to ask in a comment if you have troubles with any of them.

  - Commit messages are descriptive and explain the rationale for changes.

  - Commits which fix bugs include `Fixes #11111` at the bottom of the commit message. If this is your first pull request and you forgot to do this, write the same statement into this text field with the pull request description.

  - New unit tests have been added for relevant changes

  - you have read the QGIS Developer Guide (https://docs.qgis.org/testing/en/docs/developers_guide/index.html) and your PR complies with its QGIS Coding Standards
-->