@NathanW2 2025-09-13
@dvdkon neat!  That PR to SIP is what I found a few weeks back when I was trying to workout why SIP was so slow.  All those nested loops are killer. Glad you have managed to fix it with some indexes.