@nirvn 2025-09-13
@dvdkon , super nice to see some progress on that front! The before time is with the slow sip 6 to begin with I assume? Do you have before times for sip4?