@dvdkon 2025-09-13
>super nice to see some progress on that front! The before time is with the slow sip 6 to begin with I assume? Do you have before times for sip4?

Thanks! I couldn't get SIP 4 working at first, presumably because NixOS' PyQt is too new for SIP 4. I set up a Debian 12 container and thankfully that worked, after some digging. For me, a clean build (and presumably everything else) with SIP 4 took 111 seconds. So still somewhat faster.

Interestingly enough, I don't think it's all about the bindings generation - the SIP 4-generated code is also faster to build. It has roughly the same amount of code (1.7M vs 1.85M), so I don't really know why. For future reference, the codegen:compilation split for SIP 4 is 38:68 s and for SIP 6 it's 66:99 s. Take the specific numbers with a grain of salt, though, this is all on a laptop without any attempt to control its power level or temperature.