@dvdkon 2025-09-14
To appease the tests, I added a copyright header to `qgis_sip_project.py` (containing mostly the previous contents of `project.py.in`), crediting the original author (@manisandro). Hopefully this is okay, since the project as a whole was licenced as GNU GPLv2+ at the time the code was written.