@nyalldawson 2025-09-14
@dvdkon 
> Hopefully this is okay, since the project as a whole was licenced as GNU GPLv2+ at the time the code was written.

That's correct :+1: