@nyalldawson 2025-09-14
@dvdkon 
> Interestingly enough, I don't think it's all about the bindings generation - the SIP 4-generated code is also faster to build. It has roughly the same amount of code (1.7M vs 1.85M), so I don't really know why. For future reference, the codegen:compilation split for SIP 4 is 38:68 s and for SIP 6 it's 46:72 s.

My gut feeling is that it's the general "qt6 takes longer to compile then qt5" issue -- see eg https://forum.qt.io/topic/142309/compilation-times-with-qt6, https://forum.qt.io/topic/162226/qt-6-compiles-twice-slower-that-qt-5, https://forum.qt.io/topic/157021/projekt-compile-time-with-qt-6-much-slower-than-qt-5/11, ...