@uclaros 2025-09-19
Applying the upstream SIP patch and using this branch has made a _huge_ difference! We're back at sip4 speeds at last!! <3

For the record, I had to change
```diff
- SET(SIP_CONCAT_PARTS 10) # 3D doesn't have enough .sip files to fill 15+ .cpp files
+ SET(SIP_CONCAT_PARTS 9) # 3D doesn't have enough .sip files to fill 15+ .cpp files
```
or I would get a build error that `sip_3d_ppart9.cpp` is missing on a local branch with some 3d modifications